Ensures any functions that read or write paths are annotated @MainActor (#24)
It's not clear to me if the ObservableObject functions need to be marked @MainActor. They accept a keypath which
is used to access a potentially MainActor-isolated property, but I wonder if keypaths work around that issue somehow,
as there are no compile-time or runtime warnings if @MainActor is omitted. For now the annotations are left in until I
understand if and why they may not be needed.