Fixes for missing inherited class support for navigation destination types
The way `build<T>(_ typedData: T) → AnyView` was implemented, it didn't support using classes where the `.navigationDestination()` is triggered off a base class and navigation was happening via a child class. This refactoring uses `Mirror(reflecting:)` to get the type info for a the given `typedData`, and if no matching navigation destination factory is found, it tries again for its superclass until either a match is found or there are no more superclasses to check.

This also required me to split out the check for `LocalDestinationID` since that doesn't need to be in the loop and only needs to be checked on `typedData` initially.